worker/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
requests==2.32.3
google-generativeai==0.8.4
groq==0.18.0
boto3==1.36.11
python-dotenv==1.0.1
